refactor(tamanu): hoist shared helpers into server_info module
