+++ crates/bestool/src/actions/tamanu/alertd.rs
@@ -12,6 +12,7 @@ use super::{
 	config::{TamanuConfig, load_config},
 	connection_url::ConnectionUrlBuilder,
 	find_tamanu,
+	server_info::fetch_device_key,
 };
 use crate::actions::Context;
 
@@ -349,56 +350,6 @@ async fn build_config(
 	}
 
 	Ok(daemon_config)
-}
-
-/// Fetch the Tamanu device key for canopy targets.
-///
-/// Best-effort: returns None if the DB connection fails or the row is missing.
-/// Alerts without canopy targets continue to work regardless.
-async fn fetch_device_key(database_url: &str) -> Option<String> {
-	use tracing::warn;
-
-	let (client, connection) = match tokio_postgres::connect(database_url, tokio_postgres::NoTls)
-		.await
-	{
-		Ok(c) => c,
-		Err(err) => {
-			warn!("failed to connect for deviceKey fetch: {err}");
-			return None;
-		}
-	};
-	tokio::spawn(async move {
-		if let Err(err) = connection.await {
-			warn!("deviceKey-fetch connection error: {err}");
-		}
-	});
-
-	match client
-		.query_opt(
-			"SELECT value FROM local_system_facts WHERE key = 'deviceKey'",
-			&[],
-		)
-		.await
-	{
-		Ok(Some(row)) => match row.try_get::<_, String>(0) {
-			Ok(pem) => {
-				info!("loaded deviceKey from Tamanu DB for canopy targets");
-				Some(pem)
-			}
-			Err(err) => {
-				warn!("deviceKey row not a string: {err}");
-				None
-			}
-		},
-		Ok(None) => {
-			info!("no deviceKey in local_system_facts; canopy targets unavailable");
-			None
-		}
-		Err(err) => {
-			warn!("failed to query deviceKey: {err}");
-			None
-		}
-	}
 }
 
 async fn default_dirs(root: &std::path::Path) -> Vec<String> {
