+++ crates/bestool/src/actions/tamanu/meta_ticket.rs
@@ -1,17 +1,20 @@
-use std::process::Command;
-
 use base64::{Engine as _, engine::general_purpose::URL_SAFE_NO_PAD};
 use clap::Parser;
 use miette::{IntoDiagnostic as _, Result, miette};
 use p256::pkcs8::DecodePrivateKey as _;
 use serde::Serialize;
 use sysinfo::System;
-use tracing::{debug, info, warn};
-use uuid::Uuid;
+use tracing::debug;
 
 use crate::actions::{
 	Context,
-	tamanu::{TamanuArgs, config::load_config, connection_url::ConnectionUrlBuilder, find_tamanu},
+	tamanu::{
+		TamanuArgs,
+		config::load_config,
+		connection_url::ConnectionUrlBuilder,
+		find_tamanu,
+		server_info::{detect_virtualisation, get_or_create_server_id, get_tailscale_info},
+	},
 };
 
 /// Generate a meta-ticket for this Tamanu server
@@ -103,34 +106,6 @@ pub async fn run(ctx: Context<TamanuArgs, MetaTicketArgs>) -> Result<()> {
 	Ok(())
 }
 
-async fn get_or_create_server_id(client: &tokio_postgres::Client) -> Result<String> {
-	let row = client
-		.query_opt(
-			"SELECT value FROM local_system_facts WHERE key = 'metaServerId'",
-			&[],
-		)
-		.await
-		.into_diagnostic()?;
-
-	if let Some(row) = row {
-		let id: String = row.try_get(0).into_diagnostic()?;
-		debug!(server_id = %id, "found existing metaServerId");
-		return Ok(id);
-	}
-
-	let id = Uuid::new_v4().to_string();
-	info!(server_id = %id, "generating new metaServerId");
-	client
-		.execute(
-			"INSERT INTO local_system_facts (key, value) VALUES ('metaServerId', $1)",
-			&[&id],
-		)
-		.await
-		.into_diagnostic()?;
-
-	Ok(id)
-}
-
 fn derive_public_key_pem(private_key_pem: &str) -> Result<String> {
 	use p256::elliptic_curve::pkcs8::EncodePublicKey as _;
 
@@ -141,47 +116,6 @@ fn derive_public_key_pem(private_key_pem: &str) -> Result<String> {
 		.to_public_key_pem(p256::pkcs8::LineEnding::LF)
 		.map_err(|e| miette!("failed to encode public key PEM: {e}"))?;
 	Ok(pem)
-}
-
-fn get_tailscale_info() -> (Option<String>, Option<String>) {
-	let output = match Command::new("tailscale")
-		.arg("status")
-		.arg("--json")
-		.output()
-	{
-		Ok(output) if output.status.success() => output,
-		Ok(output) => {
-			debug!(
-				status = %output.status,
-				"tailscale status command failed"
-			);
-			return (None, None);
-		}
-		Err(e) => {
-			debug!(error = %e, "tailscale not available");
-			return (None, None);
-		}
-	};
-
-	let parsed: serde_json::Value = match serde_json::from_slice(&output.stdout) {
-		Ok(v) => v,
-		Err(e) => {
-			warn!(error = %e, "failed to parse tailscale status JSON");
-			return (None, None);
-		}
-	};
-
-	let self_node = &parsed["Self"];
-	let ip = self_node["TailscaleIPs"]
-		.as_array()
-		.and_then(|ips| ips.first())
-		.and_then(|ip| ip.as_str())
-		.map(String::from);
-	let name = self_node["DNSName"]
-		.as_str()
-		.map(|s| s.trim_end_matches('.').to_string());
-
-	(ip, name)
 }
 
 fn detect_hosting() -> Option<String> {
@@ -241,102 +175,9 @@ fn is_ec2() -> bool {
 	false
 }
 
-fn detect_virtualisation() -> Option<String> {
-	let output = Command::new("systemd-detect-virt").output().ok()?;
-
-	let stdout = String::from_utf8_lossy(&output.stdout).trim().to_string();
-	if stdout.is_empty() {
-		return None;
-	}
-
-	Some(stdout)
-}
-
 #[cfg(test)]
 mod tests {
-	use std::sync::Mutex;
-
-	use tokio_postgres::NoTls;
-
 	use super::*;
-
-	static DB_TEST_MUTEX: Mutex<()> = Mutex::new(());
-
-	async fn test_db_client() -> tokio_postgres::Client {
-		let url = std::env::var("DATABASE_URL").expect("DATABASE_URL must be set for tests");
-		let (client, connection) = tokio_postgres::connect(&url, NoTls).await.unwrap();
-		tokio::spawn(async move {
-			if let Err(e) = connection.await {
-				eprintln!("connection error: {e}");
-			}
-		});
-
-		client
-			.batch_execute(
-				"CREATE TABLE IF NOT EXISTS local_system_facts (
-					key TEXT PRIMARY KEY,
-					value TEXT NOT NULL
-				)",
-			)
-			.await
-			.unwrap();
-		client
-			.execute(
-				"DELETE FROM local_system_facts WHERE key = 'metaServerId'",
-				&[],
-			)
-			.await
-			.unwrap();
-
-		client
-	}
-
-	#[tokio::test]
-	async fn test_get_or_create_server_id_generates_new() {
-		let _lock = DB_TEST_MUTEX.lock().unwrap();
-		let client = test_db_client().await;
-
-		let id = get_or_create_server_id(&client).await.unwrap();
-		assert!(!id.is_empty());
-		uuid::Uuid::parse_str(&id).expect("should be a valid UUID");
-
-		let row = client
-			.query_one(
-				"SELECT value FROM local_system_facts WHERE key = 'metaServerId'",
-				&[],
-			)
-			.await
-			.unwrap();
-		let stored: String = row.get(0);
-		assert_eq!(stored, id);
-	}
-
-	#[tokio::test]
-	async fn test_get_or_create_server_id_returns_existing() {
-		let _lock = DB_TEST_MUTEX.lock().unwrap();
-		let client = test_db_client().await;
-
-		client
-			.execute(
-				"INSERT INTO local_system_facts (key, value) VALUES ('metaServerId', 'existing-id-123')",
-				&[],
-			)
-			.await
-			.unwrap();
-
-		let id = get_or_create_server_id(&client).await.unwrap();
-		assert_eq!(id, "existing-id-123");
-	}
-
-	#[tokio::test]
-	async fn test_get_or_create_server_id_is_stable() {
-		let _lock = DB_TEST_MUTEX.lock().unwrap();
-		let client = test_db_client().await;
-
-		let id1 = get_or_create_server_id(&client).await.unwrap();
-		let id2 = get_or_create_server_id(&client).await.unwrap();
-		assert_eq!(id1, id2);
-	}
 
 	#[test]
 	fn test_derive_public_key_pem() {
